pyproject.toml
[tool.poetry]
name = "python-flask"
version = "0.1.0"
description = "Start a simple web app with the Flask framework for Python"
authors = ["Google LLC"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.1"
python-dotenv = "^1.1.1"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.10"
flask-migrate = "^4.1.0"
bcrypt = "^4.3.0"
pyjwt = "^2.10.1"
boto3 = "^1.40.15"
tzdata = "^2025.2"

[tool.poetry.scripts]
app = 'src.main:main'

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
notebook = "^7.4.5"


[tool.poetry.group.prod.dependencies]
gunicorn = "^23.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
